# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Company Search API Document
  version: 1.0.0
paths:
  /api/v1/companies:
    get:
      operationId: ListCompanies
      summary: Get and Search Companies
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: sort_type
          required: true
          schema:
            type: string
            enum:
              - net_sales
              - average_annual_salary
              - ordinary_income
        - in: query
          name: q
          description: query for company_name or security_id
          schema:
            type: string
        - in: query
          name: industry_id
          schema:
            type: integer
        - in: query
          name: market_id
          schema:
            type: integer
      responses:
        "200":
          description: get Companies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/responseCompanies"
  /api/v1/companies/{code}:
    get:
      operationId: GetCompany
      summary: Get Company
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: get Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/responseCompany"
  /api/v1/industries:
    get:
      operationId: ListIndustries
      summary: Get Industries
      responses:
        "200":
          description: get Industries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/responseIndustries"
  /api/v1/industries/{id}:
    get:
      operationId: GetIndustry
      summary: Get Industry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: get Industry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/responseIndustry"
  /api/v1/markets:
    get:
      operationId: ListMarkets
      summary: Get Markets
      responses:
        "200":
          description: get Markets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/responseMarkets"
  /api/v1/markets/{id}:
    get:
      operationId: GetMarket
      summary: Get Market
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: get Market
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/responseMarket"
  /api/v1/security_codes:
    get:
      operationId: ListSecurityCodes
      summary: Get CompanyIds
      responses:
        "200":
          description: get CompanyIds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/responseSecurityCodes"
  /api/v1/industry_ids:
    get:
      operationId: ListIndustryIds
      summary: Get IndustryIds
      responses:
        "200":
          description: get IndustryIds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/responseIndustryIDs"
  /api/v1/market_ids:
    get:
      operationId: ListMarketIds
      summary: Get MarketIds
      responses:
        "200":
          description: get MarketIds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/responseMarketIDs"
components:
  schemas:
    responseCompanies:
      type: object
      required:
        - meta
        - companies
      properties:
        meta:
          $ref: "#/components/schemas/meta"
        companies:
          type: array
          items:
            $ref: "#/components/schemas/eachCompany"
    eachCompany:
      type: object
      required:
        - security_code
        - security_name
        - security_name_en
        - net_sales
        - ordinary_income
        - average_annual_salary
        - industry_name
        - market_name
        - industry_code
        - market_id
      properties:
        security_code:
          type: integer
        security_name:
          type: string
        security_name_en:
          type: string
        net_sales:
          type: integer
          format: int64
          nullable: true
        ordinary_income:
          type: integer
          format: int64
          nullable: true
        average_annual_salary:
          type: integer
          format: int64
          nullable: true
        industry_name:
          type: string
        industry_code:
          type: integer
          format: int32
          description: 業種コード
        market_name:
          type: string
        market_id:
          type: integer
          format: int32
          description: 市場コード
    responseCompany:
      type: object
      required:
        - company
      properties:
        company:
          $ref: "#/components/schemas/Company"
    Company:
      type: object
      required:
        - security_id
        - number_of_employees
        - average_annual_salary
        - average_age_years
        - average_length_of_service_years
        - security_code
        - security_name
        - market_name
        - market_id
        - industry_name
        - industry_id
        - industry_code
        - company_name
        - company_name_en
        - head_office_location
        - submitted_at
        - representative
        - period_started_at
        - period_ended_at
        - period_ended_at_year
        - period_ended_at_month
        - net_sales
        - last_year_net_sales
        - operating_income
        - last_year_operating_income
        - ordinary_income
        - last_year_ordinary_income
        - capital_stock
        - net_assets
        - total_assets
        - equity_to_asset_ratio
        - rate_of_return_on_equity
        - price_earnings_ratio
        - net_cash_provided_by_used_in_operating_activities
        - net_cash_provided_by_used_in_investing_activities
        - net_cash_provided_by_used_in_financing_activities
        - cash_and_cash_equivalents
        - consolidated_number_of_employees
        - total_number_of_issued_shares
        - payout_ratio
      properties:
        security_id:
          type: integer
          format: int64
          description: 証券ID
        # 従業員情報
        number_of_employees:
          type: integer
          format: int64
          nullable: true
          description: 従業員数
        average_annual_salary:
          type: integer
          format: int64
          nullable: true
          description: 平均年収
        average_age_years:
          type: number
          format: double
          nullable: true
          description: 平均年齢
        average_length_of_service_years:
          type: number
          format: double
          nullable: true
          description: 平均勤続年数
        # 基本情報
        security_code:
          type: integer
          description: 銘柄コード
        security_name:
          type: string
          description: 銘柄名
        market_name:
          type: string
          description: 市場名
        market_id:
          type: integer
          format: int64
          description: 市場コード
        industry_name:
          type: string
          description: 業種名
        industry_id:
          type: integer
          format: int32
          description: 業種ID
        industry_code:
          type: integer
          format: int32
          description: 業種コード
        company_name:
          type: string
          description: 会社名
        company_name_en:
          type: string
          description: 会社名(英語)
        head_office_location:
          type: string
          description: 本店所在地
        submitted_at:
          type: string
          description: 提出日
        representative:
          type: string
          description: 代表者
        # 決算情報
        period_started_at:
          type: string
          example: 2020年4月1日
          description: 事業年度の開始日
        period_ended_at:
          type: string
          example: 2021年3月31日
          description: 事業年度の終了日
        period_ended_at_year:
          type: integer
          example: 2021
          description: 事業年度の終了年
        period_ended_at_month:
          type: integer
          example: 3
          description: 事業年度の終了月
        net_sales:
          type: integer
          format: int64
          nullable: true
          description: 売上高(当年)
        last_year_net_sales:
          type: integer
          format: int64
          nullable: true
          description: 売上高(前年)
        operating_income:
          type: integer
          format: int64
          nullable: true
          description: 営業利益(当年)
        last_year_operating_income:
          type: integer
          format: int64
          nullable: true
          description: 営業利益(前年)
        ordinary_income:
          type: integer
          format: int64
          nullable: true
          description: 経常利益(当年)
        last_year_ordinary_income:
          type: integer
          format: int64
          nullable: true
          description: 経常利益(前年)
        capital_stock:
          type: integer
          format: int64
          nullable: true
          description: 資本金
        net_assets:
          type: integer
          format: int64
          nullable: true
          description: 純資産
        total_assets:
          type: integer
          format: int64
          nullable: true
          description: 総資産
        equity_to_asset_ratio:
          type: number
          format: double
          nullable: true
          description: 自己資本比率
        rate_of_return_on_equity:
          type: number
          format: double
          nullable: true
          description: 自己資本利益率(ROE)
        price_earnings_ratio:
          type: number
          format: double
          nullable: true
          description: 株価収益率(PER)
        net_cash_provided_by_used_in_operating_activities:
          type: integer
          format: int64
          nullable: true
          description: 営業キャッシュフロー
        net_cash_provided_by_used_in_investing_activities:
          type: integer
          format: int64
          nullable: true
          description: 投資キャッシュフロー
        net_cash_provided_by_used_in_financing_activities:
          type: integer
          format: int64
          nullable: true
          description: 財務キャッシュフロー
        cash_and_cash_equivalents:
          type: integer
          format: int64
          nullable: true
          description: 現金及び現金同等物
        consolidated_number_of_employees:
          type: integer
          format: int64
          nullable: true
          description: 連結従業員数
        total_number_of_issued_shares:
          type: integer
          format: int64
          nullable: true
          description: 発行済株式総数
        payout_ratio:
          type: number
          format: double
          nullable: true
          description: 配当性向
    responseIndustries:
      type: object
      required:
        - industry_categories
      properties:
        industry_categories:
          type: array
          items:
            $ref: "#/components/schemas/eachIndustryCategory"
    eachIndustryCategory:
      type: object
      required:
        - id
        - name
        - industries
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        industries:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/eachIndustryCategoryIndustry"
    eachIndustryCategoryIndustry:
      type: object
      required:
        - id
        - name
        - code
        - count
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        code:
          type: integer
          format: int32
          description: 33業種コード
        count:
          type: integer
          format: int32
          description: 業種に属する企業数
    responseIndustryIDs:
      type: object
      required:
        - industry_ids
      properties:
        industry_ids:
          type: array
          items:
            type: integer
            format: int64
    responseIndustry:
      type: object
      required:
        - industry
      properties:
        industry:
          $ref: "#/components/schemas/industry"
    industry:
      type: object
      required:
        - id
        - name
        - code
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        code:
          type: integer
    responseMarketIDs:
      type: object
      required:
        - market_ids
      properties:
        market_ids:
          type: array
          items:
            type: integer
            format: int64
    responseMarkets:
      type: object
      required:
        - markets
      properties:
        markets:
          type: array
          items:
            $ref: "#/components/schemas/eachMarket"
    eachMarket:
      type: object
      required:
        - id
        - name
        - count
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        count:
          type: integer
          description: 市場に属する企業数
    responseMarket:
      type: object
      required:
        - market
      properties:
        market:
          $ref: "#/components/schemas/market"
    market:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    responseSecurityCodes:
      type: object
      required:
        - security_codes
      properties:
        security_codes:
          type: array
          items:
            type: integer
            format: int64
    meta:
      type: object
      required:
        - offset_count
        - total_count
        - limit_count
        - current_page
        - last_page
        - prev_page
        - next_page
        - sort_type
      properties:
        offset_count:
          type: integer
          description: アイテムの取得開始位置
        total_count:
          type: integer
          description: 総アイテム数
        limit_count:
          type: integer
          description: 1ページあたりのアイテム数
        current_page:
          type: integer
          description: 現在のページ番号
        last_page:
          type: integer
          description: 最後ののページ番号
        prev_page:
          type: integer
          nullable: true
          description: 前のページ番号
        next_page:
          type: integer
          nullable: true
          description: 次のページ番号
        sort_type:
          type: string
          description: ソートタイプ